+++ automatedtests/src/androidTestR2_13_1/java/com/google/android/exoplayer2/ext/ima/ImaUtil.java
@@ -16,6 +16,7 @@
 package com.google.android.exoplayer2.ext.ima;
 
 import android.content.Context;
+import android.os.Looper;
 import android.view.View;
 import android.view.ViewGroup;
 import androidx.annotation.Nullable;
@@ -33,8 +34,8 @@ import com.google.ads.interactivemedia.v3.api.FriendlyObstructionPurpose;
 import com.google.ads.interactivemedia.v3.api.ImaSdkSettings;
 import com.google.ads.interactivemedia.v3.api.UiElement;
 import com.google.ads.interactivemedia.v3.api.player.VideoAdPlayer;
+import com.google.ads.interactivemedia.v3.api.player.VideoProgressUpdate;
 import com.google.android.exoplayer2.C;
-import com.google.android.exoplayer2.source.ads.AdPlaybackState;
 import com.google.android.exoplayer2.source.ads.AdsLoader.OverlayInfo;
 import com.google.android.exoplayer2.upstream.DataSchemeDataSource;
 import com.google.android.exoplayer2.upstream.DataSpec;
@@ -110,6 +111,8 @@ import java.util.Set;
     public final boolean playAdBeforeStartPosition;
     public final int mediaBitrate;
     @Nullable
+    public final Boolean enableContinuousPlayback;
+    @Nullable
     public final List<String> adMediaMimeTypes;
     @Nullable
     public final Set<UiElement> adUiElements;
@@ -132,6 +135,7 @@ import java.util.Set;
         boolean focusSkipButtonWhenAvailable,
         boolean playAdBeforeStartPosition,
         int mediaBitrate,
+        @Nullable Boolean enableContinuousPlayback,
         @Nullable List<String> adMediaMimeTypes,
         @Nullable Set<UiElement> adUiElements,
         @Nullable Collection<CompanionAdSlot> companionAdSlots,
@@ -146,6 +150,7 @@ import java.util.Set;
       this.focusSkipButtonWhenAvailable = focusSkipButtonWhenAvailable;
       this.playAdBeforeStartPosition = playAdBeforeStartPosition;
       this.mediaBitrate = mediaBitrate;
+      this.enableContinuousPlayback = enableContinuousPlayback;
       this.adMediaMimeTypes = adMediaMimeTypes;
       this.adUiElements = adUiElements;
       this.companionAdSlots = companionAdSlots;
@@ -156,6 +161,9 @@ import java.util.Set;
       this.debugModeEnabled = debugModeEnabled;
     }
   }
+
+  public static final int TIMEOUT_UNSET = -1;
+  public static final int BITRATE_UNSET = -1;
 
   /**
    * Returns the IMA {@link FriendlyObstructionPurpose} corresponding to the given {@link
@@ -177,15 +185,14 @@ import java.util.Set;
   }
 
   /**
-   * Returns an initial {@link AdPlaybackState} with ad groups at the provided {@code cuePoints}.
+   * Returns the microsecond ad group timestamps corresponding to the specified cue points.
    *
-   * @param cuePoints The cue points of the ads in seconds.
-   * @return The {@link AdPlaybackState}.
+   * @param cuePoints The cue points of the ads in seconds, provided by the IMA SDK.
+   * @return The corresponding microsecond ad group timestamps.
    */
-  public static AdPlaybackState getInitialAdPlaybackStateForCuePoints(List<Float> cuePoints) {
+  public static long[] getAdGroupTimesUsForCuePoints(List<Float> cuePoints) {
     if (cuePoints.isEmpty()) {
-      // If no cue points are specified, there is a preroll ad.
-      return new AdPlaybackState(/* adGroupTimesUs...= */ 0);
+      return new long[]{0L};
     }
 
     int count = cuePoints.size();
@@ -201,7 +208,7 @@ import java.util.Set;
     }
     // Cue points may be out of order, so sort them.
     Arrays.sort(adGroupTimesUs, 0, adGroupIndex);
-    return new AdPlaybackState(adGroupTimesUs);
+    return adGroupTimesUs;
   }
 
   /**
@@ -232,6 +239,28 @@ import java.util.Set;
     // a single ad, ad group or the whole timeline.
     return adError.getErrorCode() == AdError.AdErrorCode.VAST_LINEAR_ASSET_MISMATCH
         || adError.getErrorCode() == AdError.AdErrorCode.UNKNOWN_ERROR;
+  }
+
+  /**
+   * Returns the looper on which all IMA SDK interaction must occur.
+   */
+  public static Looper getImaLooper() {
+    // IMA SDK callbacks occur on the main thread. This method can be used to check that the player
+    // is using the same looper, to ensure all interaction with this class is on the main thread.
+    return Looper.getMainLooper();
+  }
+
+  /**
+   * Returns a human-readable representation of a video progress update.
+   */
+  public static String getStringForVideoProgressUpdate(VideoProgressUpdate videoProgressUpdate) {
+    if (VideoProgressUpdate.VIDEO_TIME_NOT_READY.equals(videoProgressUpdate)) {
+      return "not ready";
+    } else {
+      return Util.formatInvariant(
+          "%d ms of %d ms",
+          videoProgressUpdate.getCurrentTimeMs(), videoProgressUpdate.getDurationMs());
+    }
   }
 
   private ImaUtil() {
